# openapi.yaml
openapi: 3.0.4
info:
  title: 'PhotoBank.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
  version: '1.0'
servers:
  - url: /api
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      operationId: Auth_Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
            text/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/register:
    post:
      tags:
        - Auth
      operationId: Auth_Register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/user:
    get:
      tags:
        - Auth
      operationId: Auth_GetUser
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserDto'
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
            text/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    put:
      tags:
        - Auth
      operationId: Auth_UpdateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/telegram/exchange:
    post:
      tags:
        - Auth
      operationId: Auth_TelegramExchange
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramExchangeRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TelegramExchangeRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TelegramExchangeRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TelegramExchangeResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramExchangeResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/TelegramExchangeResponse'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    LoginRequestDto:
      required:
        - email
        - password
        - rememberMe
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        rememberMe:
          type: boolean
      additionalProperties: false
    LoginResponseDto:
      required:
        - token
      type: object
      properties:
        token:
          type: string
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    RegisterRequestDto:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
...
    TelegramExchangeRequest:
      type: object
      properties:
        telegramUserId:
          type: integer
          format: int64
        username:
          type: string
          nullable: true
      additionalProperties: false
    TelegramExchangeResponse:
      type: object
      properties:
        accessToken:
          type: string
          nullable: true
        expiresIn:
          type: integer
          format: int32
      additionalProperties: false
...
    UpdateUserDto:
      type: object
      properties:
        phoneNumber:
          type: string
          nullable: true
        telegramUserId:
          type: integer
          format: int64
          nullable: true
        telegramSendTimeUtc:
          type: string
          format: date-span
          nullable: true
      additionalProperties: false
    UserDto:
      required:
        - email
      type: object
      properties:
        email:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        telegramUserId:
          type: integer
          format: int64
          nullable: true
        telegramSendTimeUtc:
          type: string
          format: date-span
          nullable: true
      additionalProperties: false
...